Port to RustCrypto
According to https://rustsec.org/advisories/RUSTSEC-2016-0005.html,
rust-crypto is unmaintained.

Crates depending on rust-crypto should be ported to other crates. One
recommendation are the crates from https://github.com/RustCrypto.

This port replaces rust-crypto with those crates, while keeping the API
of rust-jwt unchanged as far as it is possible.

Unfortunately, rust-jwt's API contains types from rust-crypto, so
changing the underlying implementations necessarily causes a breaking
change.

However, as can be seen in the examples, existing code only needs to
change the types used, not the actual code.

Before publishing a release containing this commit, the version number
of rust-jwt should be updated to reflect the breaking change.